+++ src/inject/hooks.rs
@@ -7,7 +7,7 @@ use super::{Inject, Result};
 #[async_trait]
 pub trait Hook: Any + Send {
     /// Handle the event by operating on the Inject container
-    async fn handle(&mut self, i: &mut Inject) -> Result<()>;
+    async fn handle(&self, i: &mut Inject) -> Result<()>;
 }
 
 /// A no-op hook that does nothing, for use as a default
@@ -15,7 +15,17 @@ pub struct NoOpHook {}
 
 #[async_trait]
 impl Hook for NoOpHook {
-    async fn handle(&mut self, _i: &mut Inject) -> Result<()> {
+    async fn handle(&self, _i: &mut Inject) -> Result<()> {
         Ok(())
+    }
+}
+
+impl Inject {
+    /// Use a Provider function to inject a dependency.
+    pub async fn handle<H>(&mut self, hook: H) -> Result<()>
+    where
+        H: Hook,
+    {
+        hook.handle(self).await
     }
 }